Rename infobox_scraper module to meta_infobox for consistency
